app/requirements.txt
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.24.25
botocore==1.27.25
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications-hq==1.7.0
django-oauth-toolkit==2.0.0
django-rest-auth==0.9.5
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djangotoolbox==1.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-api-logger==1.1.10
drf-yasg==1.20.0
easyocr==1.4.2
environ==1.0
facebook-sdk==3.1.0
Faker==13.12.1
flake8==4.0.1
fonttools==4.33.3
fuzzywuzzy==0.18.0
google-api-core==2.8.1
google-api-python-client==2.47.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imageio==2.19.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jsonschema==3.2.0
jwcrypto==1.3.1
kiwisolver==1.4.3
Markdown==3.3.4
MarkupSafe==2.1.1
martor==1.6.13
matplotlib==3.5.2
mccabe==0.6.1
mongoengine==0.24.1
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.2
paramiko==2.11.0
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==1.28.4
pyrsistent==0.18.1
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
swapper==1.3.0
texttable==1.6.4
tifffile==2022.5.4
tika==1.24
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
Wand==0.6.7
webencodings==0.5.1
websocket-client==0.59.0
wrapt==1.14.1
channels==3.0.4
djangochannelsrestframework==1.0.0
selenium==4.3.0
channels-redis==3.4.0